pages: moves everything to `pocs` folder and use `.public` suffix
Also remove "lorem ipsum" dependencies and upgrade the other ones
