+++ pages/pocs/home-09.public.js
@@ -1,10 +1,15 @@
-import React, { createContext, useContext, useState, useCallback, useEffect } from 'react'
+import React, {
+  createContext,
+  useContext,
+  useState,
+  useCallback,
+  useEffect,
+} from "react";
 import { CgTab } from "react-icons/cg";
 import { MdAccountCircle } from "react-icons/md";
 import { MdMonetizationOn } from "react-icons/md";
 import Link from "next/link";
 import { FaStar } from "react-icons/fa";
-import { LoremIpsum } from "lorem-ipsum";
 
 export const newsItems = [
   {
@@ -136,46 +141,55 @@ export const newsItems = [
   },
 ];
 
-const ThemeModeContext = createContext({})
+const ThemeModeContext = createContext({});
 
 const ThemeModeProvider = ({ children }) => {
-  const [isDarkModeOn, setIsDarkModeOn] = useState(false)
+  const [isDarkModeOn, setIsDarkModeOn] = useState(false);
 
   useEffect(() => {
-    const htmlELement = document.querySelector('html')
-    const themeMode = localStorage.getItem('@tabnews:theme')
+    const htmlELement = document.querySelector("html");
+    const themeMode = localStorage.getItem("@tabnews:theme");
 
     if (themeMode) {
-      JSON.parse(themeMode.toLowerCase()) ? htmlELement.classList.add("dark") : htmlELement.classList.remove("dark");
-      JSON.parse(themeMode.toLowerCase()) ? setIsDarkModeOn(true) : setIsDarkModeOn(false)
+      JSON.parse(themeMode.toLowerCase())
+        ? htmlELement.classList.add("dark")
+        : htmlELement.classList.remove("dark");
+      JSON.parse(themeMode.toLowerCase())
+        ? setIsDarkModeOn(true)
+        : setIsDarkModeOn(false);
     } else {
-      localStorage.setItem('@tabnews:theme', false)
+      localStorage.setItem("@tabnews:theme", false);
     }
-  }, [isDarkModeOn])
+  }, [isDarkModeOn]);
 
-  const handleDarkModeChange = useCallback(()=> {
-    localStorage.setItem('@tabnews:theme', !isDarkModeOn)
-    setIsDarkModeOn(!isDarkModeOn)
-    return
-  },[isDarkModeOn])
+  const handleDarkModeChange = useCallback(() => {
+    localStorage.setItem("@tabnews:theme", !isDarkModeOn);
+    setIsDarkModeOn(!isDarkModeOn);
+    return;
+  }, [isDarkModeOn]);
 
   return (
     <ThemeModeContext.Provider value={{ isDarkModeOn, handleDarkModeChange }}>
       {children}
     </ThemeModeContext.Provider>
-  )
-}
+  );
+};
 
 export function useTheme() {
-  return useContext(ThemeModeContext)
+  return useContext(ThemeModeContext);
 }
 
-
 export function ToggleButton() {
-  const {isDarkModeOn, handleDarkModeChange} = useTheme()
+  const { isDarkModeOn, handleDarkModeChange } = useTheme();
 
   return (
-    <button className="border border-gray-300 dark:border-darkTheme-secondary text-gray-700 focus:outline-none px-4 rounded-lg dark:text-darkTheme-primary" type="button" onClick={()=>{handleDarkModeChange()}}>
+    <button
+      className="border border-gray-300 dark:border-darkTheme-secondary text-gray-700 focus:outline-none px-4 rounded-lg dark:text-darkTheme-primary"
+      type="button"
+      onClick={() => {
+        handleDarkModeChange();
+      }}
+    >
       {isDarkModeOn ? "Desligar Dark Mode" : "Ligar Dark Mode"}
     </button>
   );
@@ -187,16 +201,20 @@ export function Header() {
       <nav className="flex items-center justify-between pt-2 pb-2 mb-3 border-b border-gray-200 dark:border-darkTheme-secondary">
         <div className="flex items-center space-x-1 text-gray-800">
           <CgTab className="w-7 h-7 dark:text-darkTheme-primary" />
-          <span className="text-md font-medium dark:text-darkTheme-primary">TabNews</span>
+          <span className="text-md font-medium dark:text-darkTheme-primary">
+            TabNews
+          </span>
         </div>
         <div className="flex space-x-2">
-          <ToggleButton/>
+          <ToggleButton />
           <div className="flex items-center pt-1 pb-1 pl-2 pr-2 font-mono text-lg text-gray-500 border border-gray-300 rounded-lg dark:text-darkTheme-primary dark:border-darkTheme-secondary">
-            <MdMonetizationOn className="w-6 h-6 mr-1 text-yellow-400 dark:text-darkTheme-yellow-400" /> 0052
-            </div>
+            <MdMonetizationOn className="w-6 h-6 mr-1 text-yellow-400 dark:text-darkTheme-yellow-400" />{" "}
+            0052
+          </div>
           <div className="flex items-center pt-1 pb-1 pl-2 pr-2 font-mono text-lg text-gray-500 border border-gray-300 rounded-lg dark:border-darkTheme-secondary dark:text-darkTheme-primary">
-            <MdAccountCircle className="w-6 h-6  mr-1 text-blue-500 dark:text-darkTheme-green-200" /> 1430
-            </div>
+            <MdAccountCircle className="w-6 h-6  mr-1 text-blue-500 dark:text-darkTheme-green-200" />{" "}
+            1430
+          </div>
         </div>
       </nav>
     </header>
@@ -204,18 +222,6 @@ export function Header() {
 }
 
 export function NewsList() {
-
-  const lorem = new LoremIpsum({
-    sentencesPerParagraph: {
-      max: 8,
-      min: 4,
-    },
-    wordsPerSentence: {
-      max: 16,
-      min: 4,
-    },
-  });
-
   return (
     <div className="flex flex-col space-y-5">
       {newsItems.map((news, index) => {
@@ -224,7 +230,10 @@ export function NewsList() {
         }
 
         return (
-          <div className="flex font-sans items-center justify-center" key={index}>
+          <div
+            className="flex font-sans items-center justify-center"
+            key={index}
+          >
             <Link href="#">
               <a className="flex flex-col items-center self-start mt-1 mr-2">
                 <div
@@ -233,7 +242,7 @@ export function NewsList() {
                     index === 0
                       ? "border-yellow-300 text-gray-900 dark:text-darkTheme-yellow-400 dark:border-darkTheme-yellow-400"
                       : "border-gray-300 text-gray-700 dark:text-darkTheme-primary dark:border-darkTheme-primary"
-                    }
+                  }
                   `}
                 >
                   {index + 1}
@@ -244,7 +253,7 @@ export function NewsList() {
                     index === 0
                       ? "bg-yellow-300 text-yellow-900 dark:text-darkTheme-background dark:bg-darkTheme-yellow-400"
                       : "bg-gray-300 text-gray-700 dark:text-darkTheme-background dark:bg-darkTheme-primary"
-                    }
+                  }
                   `}
                 >
                   {news.coins}
@@ -253,16 +262,18 @@ export function NewsList() {
             </Link>
             <div className="flex-1 overflow-hidden">
               <Link href="/post-01">
-                <a className={`inline-block text-base text-lg font-semibold leading-snug text-gray-900 ${
+                <a
+                  className={`inline-block text-base text-lg font-semibold leading-snug text-gray-900 ${
                     index === 0
                       ? "dark:text-darkTheme-yellow-400"
                       : "dark:text-darkTheme-primary"
-                    }
-                  `}>
+                  }
+                  `}
+                >
                   {news.title}
                 </a>
               </Link>
-              
+
               <div className="text-md tracking-tight text-gray-400 dark:text-darkTheme-yellow-200">
                 {news.comment && (
                   <div className="flex items-center">
@@ -287,7 +298,6 @@ export function NewsList() {
     </div>
   );
 }
-
 
 export default function Home() {
   return (
